requirements.txt
# pip install -r requirements.txt

pyserial~=3.5
